+++ CHANGELOG.md
@@ -32,9 +32,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dashboard v2.4.0
 - Ansible Compatibility "2.0"
 
-## [1.1.2] - 2023-06-29
+## [1.1.2] - 2023-08-28
 ### Changed
-- -n option argument to set number of nodes (default=1).
+- -n option argument to set or disable the number of nodes (default=1).
+- `KUBECONFIG` environment variable.
 - Ubuntu 22.04
 - Dashboard v2.7.0
 - Weave CNI v2.8.1